rename site to site_metadata
