+++ .gencode_hash.txt
@@ -15,14 +15,14 @@ c601733d04b0d389d7800a801b0862da41567d1f1fa63d9529a462b8269f0462  gencode/docs/e
 9ae793a641f2e053c5af734bf813cd2637a675330acb8d9e903ef03b4286e007  gencode/docs/metadata.html
 9300480de06d728cfe366304b4a06ad1fc135cd27fa5fd3f1bc59e33703044a5  gencode/docs/persistent_device.html
 5d039d607af9ec75ee552dfe36b16c702687ea16f5663f41fc49b4533b86e00d  gencode/docs/properties.html
-15e84a23db290ff0aaa378884c8cb520f38e925171badffad7068fc1ccca5eed  gencode/docs/readme.md
+4b867e67badf5c4c5d29266fa7e5314daecf53cc59f20f03ec16943ca8d6d227  gencode/docs/readme.md
 f46367beaab5886ef87ce70b542611134c3565daa53e367de4f49dab51214f12  gencode/docs/reflect_config.html
 9332e44c87dae9261b079424e748d5ee1df08a3c6b39987d254ebf78274e2f34  gencode/docs/reflect_state.html
 741b880216be3743f6747800a042f2dbd89f3b0344c6b0a965f4bc010f03a930  gencode/docs/schema_doc.css
 878ea88206c974f40643c3cc430875f9c4e8c5e3fd6bcd6358bd3eb6d48699a9  gencode/docs/schema_doc.min.js
 7ed934930aee763e0beebc349725ba3909115e8d346bb762f28bcbe745bb163a  gencode/docs/schema_extras.js
-c23f763d6f767422f80ff9ed13a1dc8d9fea7fd0be54f75022907b8a37c3f8cb  gencode/docs/site.html
 09988222bc4e7cda512a3a59a3af1952a93546b37b2d39f25fd754a4f17bb79e  gencode/docs/site_cloud.html
+c23f763d6f767422f80ff9ed13a1dc8d9fea7fd0be54f75022907b8a37c3f8cb  gencode/docs/site_metadata.html
 63cd94346f634ca2945982c147546cb1103fcfea4c42eac04b065ca838f5217f  gencode/docs/state.html
 f10172c9b4cd350cae1f673077b07b1362168599b966a78d4ec4ef19b7bbf441  gencode/docs/state_mapping.html
 552f85739f547960d2c1b96c99152ff831a036ce4d19908a5361b35a2d5f2132  gencode/docs/state_validation.html
@@ -95,7 +95,7 @@ abe99dd74122c186403baa6982300a9d5968f8bbb7a67b1689104111b98f32fb  gencode/java/u
 362499688dd17f0596e5d6f064db0fe3bb3f3a25bdf01319624ba423de779220  gencode/java/udmi/schema/SetupReflectorConfig.java
 649c0291ad81421fc51da0c2f7da3286628127157a5a6eef77610e8c37c14941  gencode/java/udmi/schema/SetupReflectorState.java
 d796835b43ee7600140544f092379c884d5c62ec9c1d4894f2c8c0f6072072e7  gencode/java/udmi/schema/Site.java
-576a8cc4e6ee120d86f7d82866e4581ecc299a9a4ab70f9c8c2159bccd64b890  gencode/java/udmi/schema/SiteLocation.java
+62122ed369638d8f8da0886cbab9f5ab7cdde8b11fcd05de20048cc533a292be  gencode/java/udmi/schema/SiteLocation.java
 580df660dad1b97399002271716d597f72aa1a6110a49de9e162104c231752f4  gencode/java/udmi/schema/State.java
 47ad54328160a1aabce719a3263d4bfea903e8e8255e04a54ace86eb095b0f0c  gencode/java/udmi/schema/StateSystemHardware.java
 c9b920d355e5f6350fa6eb978afbfe55dd1438df171d2ea215845d5ac0b3e7e9  gencode/java/udmi/schema/StateSystemOperation.java
@@ -112,7 +112,7 @@ ac6f8fd87c8986cce01e872460c15ff6fe71e3816f9bde610acfe25f7d38c8d4  gencode/java/u
 f7d117dc8b9764acf0c95a13a2bfdfbdf31d1a8ec83a707448aa4d7391ef07e2  gencode/java/udmi/schema/ValidationState.java
 e007ddd1ceeae3603c85110c33e1bb4a418ff9c7a791ca0df25b7ea3caeafd36  gencode/java/udmi/schema/ValidationSummary.java
 b77d953fd22e655c0f10ae32deeaa222769d971f8c38b3379eba45720fb910cc  gencode/java/udmi/schema/VirtualEquipmentLinks.java
-12834c04564736e2e5651ab3387c62dbbaa916ec18f5fa663f55536b1ebda7f6  gencode/python/udmi/schema/__init__.py
+de8021495fb6c645bff0408a403a20a0a52c1d531a448243c04d59134df853f3  gencode/python/udmi/schema/__init__.py
 4b25dd95f863059b761269f93adcae7049507924a1c6e74d6856849203c179db  gencode/python/udmi/schema/ancillary_properties.py
 dab4f5fca272ec48c2881bca2b6bc43786ada47fa1f6dd935c35f7ce0eb6b0f6  gencode/python/udmi/schema/building_translation.py
 470b688984b89b25fcdfa8e08bd95b0c5d8c551d53a6ab5512503ee39419e6fa  gencode/python/udmi/schema/category.py
@@ -167,10 +167,10 @@ aafe6e70c281152db958adf77a024e3e9fab8293927106297c5ec48c11f54e27  gencode/python
 a58f8c98e837a5b56126ca0f410e02f1e9cfcd80a8cb429e0ef522defab1f690  gencode/python/udmi/schema/properties.py
 5dad0f2ee3e94ec32d11d1de3bfde35d087d3561ee18c0a9eeb0722d17d0f58a  gencode/python/udmi/schema/reflect_config.py
 1a4eef286957d77418777eb7f4cd1ee13aa24c2d916ec71a7ff4e56de76e303e  gencode/python/udmi/schema/reflect_state.py
-b593339fda1fd8ae54913baefadb0f43233efda433aef187a081317ea8d91e28  gencode/python/udmi/schema/site.py
 a3ca15fbbd6423971005b5caedf8633eb004945f7ef957ce8003d256234d899a  gencode/python/udmi/schema/site_cloud.py
-f5cc053a9ae343af8af566b06ba037301258277a0b8324b3e5e0037c1f294485  gencode/python/udmi/schema/site_location.py
-ac169c09da87014a968dade8b5a7ab4725c26dffb92f2394a79bd658500a9934  gencode/python/udmi/schema/site_origin.py
+339832e11f6e8d9a78871aa013d87c513275178ca3d53f68b2c6bceff78acfda  gencode/python/udmi/schema/site_metadata.py
+f0e3e425a582c9fa77640069777d70ee7adac5581be347685629d850cf3f083e  gencode/python/udmi/schema/site_metadata_location.py
+fdaa7e2dcd7aba3766848d9bfcf13107962260945e943a7ac784a0d661e3ce53  gencode/python/udmi/schema/site_metadata_origin.py
 32bc70a30e37e89cfae14b44add18d546a6f9e00a3ec3519ede9c7486114d39c  gencode/python/udmi/schema/state.py
 4a908cee3fb8afb559bcbfa594e57dbc515a35e4468e02600751b2fcce05a238  gencode/python/udmi/schema/state_blobset.py
 b7f07c4a74dcd5262cb3eb817cc568f73efd253c3ae5be44620a0204507105a6  gencode/python/udmi/schema/state_blobset_blob.py
