+++ gencode/python/udmi/schema/site_metadata_location.py
@@ -1,24 +1,22 @@
-"""Generated class for site_origin.json"""
+"""Generated class for site_metadata_location.json"""
 
 
 class SiteLocation:
   """Generated schema class"""
 
   def __init__(self):
+    self.address = None
     self.lat = None
     self.long = None
-    self.alt = None
-    self.orientation = None
 
   @staticmethod
   def from_dict(source):
     if not source:
       return None
     result = SiteLocation()
+    result.address = source.get('address')
     result.lat = source.get('lat')
     result.long = source.get('long')
-    result.alt = source.get('alt')
-    result.orientation = source.get('orientation')
     return result
 
   @staticmethod
@@ -39,12 +37,10 @@ class SiteLocation:
 
   def to_dict(self):
     result = {}
+    if self.address:
+      result['address'] = self.address # 5
     if self.lat:
       result['lat'] = self.lat # 5
     if self.long:
       result['long'] = self.long # 5
-    if self.alt:
-      result['alt'] = self.alt # 5
-    if self.orientation:
-      result['orientation'] = self.orientation # 5
     return result
